+++ packages-tooling/plugin-conventions/src/hmr.ts
@@ -99,7 +99,6 @@ export const getHmrCode = (className: string, moduleText?: string, moduleNames?:
           controller.hE(hydrationInst, hydrationContext);
         }
         h.parentNode.replaceChild(controller.host, h);
-        controller.hostController = null;
         controller.deactivate(controller, controller.parent ?? null, 0);
         controller.activate(controller, controller.parent ?? null, 0);
       });