+++ packages/__tests__/src/3-runtime-html/custom-elements.spec.ts
@@ -1,7 +1,22 @@
-import { AppTask, Aurelia, bindable, BindingMode, customElement, CustomElement, IAppRoot, IAurelia, IKeyMapping, ShortHandBindingSyntax, ValueConverter } from '@aurelia/runtime-html';
+import {
+  AppTask,
+  Aurelia,
+  bindable,
+  BindingMode,
+  Controller,
+  customElement,
+  CustomElement,
+  CustomElementDefinition,
+  IAppRoot,
+  IAurelia,
+  ICustomElementController,
+  IKeyMapping,
+  ShortHandBindingSyntax,
+  ValueConverter,
+} from '@aurelia/runtime-html';
 import { assert, createFixture } from '@aurelia/testing';
 import { delegateSyntax } from '@aurelia/compat-v1';
-import { resolve } from '@aurelia/kernel';
+import { IContainer, resolve } from '@aurelia/kernel';
 import { IObserverLocator, observable } from '@aurelia/runtime';
 
 describe('3-runtime-html/custom-elements.spec.ts', function () {
@@ -679,6 +694,90 @@ describe('3-runtime-html/custom-elements.spec.ts', function () {
     });
   });
 
+  describe('ref', function () {
+    it('updates ref when key changes', function () {
+      const { component } = createFixture(
+        `<div ref="$this[key]"></div>`,
+        class {
+          key = 'a';
+          a: HTMLElement;
+          b: HTMLElement;
+        },
+      );
+
+      assert.strictEqual(component.a?.tagName, 'DIV');
+      component.key = 'b';
+      assert.deepEqual(
+        [component.a?.tagName, component.b?.tagName],
+        // ref binding is not responsible for cleaning up if the key has been changed.
+        // It's also not possible to clean up if the key has been changed.
+        ['DIV', 'DIV']
+      );
+    });
+
+    it('sets ref to null when unbound', function () {
+      const { component } = createFixture(
+        `<div if.bind="show" ref="el">`,
+        class {
+          show = true;
+          el: HTMLElement | 1;
+        }
+      );
+
+      assert.strictEqual((component.el as HTMLElement)?.tagName, 'DIV');
+      component.show = false;
+      assert.strictEqual(component.el, null);
+    });
+
+    it('does not update ref if the value has been changed before unbind', function () {
+      const { component } = createFixture(
+        `<div if.bind="show" ref="el">`,
+        class {
+          show = true;
+          el: HTMLElement | 1;
+        }
+      );
+
+      assert.strictEqual((component.el as HTMLElement)?.tagName, 'DIV');
+      component.el = 1;
+      component.show = false;
+      assert.strictEqual(component.el, 1);
+    });
+
+    it('does not throw when binding to a null object', function () {
+      const { component } = createFixture(
+        `<div if.bind="show" ref="els.div">`,
+        class {
+          els: Record<string, HTMLElement> | null;
+        }
+      );
+
+      assert.equal(component.els, null);
+    });
+
+    it('updates ref when property key changes GH #2106', function () {
+      const { component, assertHtml, flush } = createFixture(
+        `<div repeat.for="item of items"
+          data-id.bind="item.id"
+          ref="children[$index]">\${item.id}</div>`,
+        class {
+          items = [{ id: 1 }];
+          children: HTMLElement[] = [];
+          add() {
+            this.items.unshift({ id: 2 });
+          }
+        }
+      );
+
+      assertHtml('<div data-id="1">1</div>', { compact: true });
+      assert.deepEqual(component.children.map(c => c.getAttribute('data-id')), ['1']);
+      component.add();
+      flush();
+      assertHtml('<div data-id="2">2</div><div data-id="1">1</div>', { compact: true });
+      assert.deepEqual(component.children.map(c => c.getAttribute('data-id')), ['2', '1']);
+    });
+  });
+
   describe('resolve', function () {
     afterEach(function () {
       assert.throws(() => resolve(class Abc {}));
@@ -1218,5 +1317,31 @@ describe('3-runtime-html/custom-elements.spec.ts', function () {
           5: { newValue: 4, oldValue: 2 }
         });
     });
+  });
+
+  it('throws when trying to create a custom element with a node associated with another element', function () {
+
+    let i = 0;
+    class App {
+      ctn = resolve(IContainer);
+      $controller: ICustomElementController;
+
+      attaching() {
+        i = 1;
+      }
+
+      attached() {
+        Controller.$el(
+          this.ctn,
+          {},
+          this.$controller.host,
+          null,
+          CustomElementDefinition.create({ name: 'abc' }),
+        );
+      }
+    }
+
+    assert.throws(() => createFixture('', App));
+    assert.strictEqual(i, 1);
   });
 });