+++ packages/router-lite/src/instructions.ts
@@ -18,13 +18,15 @@ import {
 import { IRouteViewModel } from './component-agent';
 import { RouteType } from './route';
 import { type $RecognizedRoute, IRouteContext, RouteContext } from './route-context';
-import { expectType, isPartialViewportInstruction, shallowEquals } from './validation';
+import { expectType, isPartialCustomElementDefinition, isPartialViewportInstruction, shallowEquals } from './validation';
 import { INavigationOptions, NavigationOptions, type RouterOptions } from './options';
 import { RouteExpression } from './route-expression';
 import { mergeURLSearchParams, tryStringify } from './util';
 import { Events, getMessage } from './events';
 import { State } from './viewport-agent';
 import { IUrlParser } from './url-parser';
+import { RouteNode } from './route-tree';
+import { RecognizedRoute } from '@aurelia/route-recognizer';
 
 export const defaultViewportName = 'default';
 export type RouteContextLike = IRouteContext | ICustomElementViewModel | ICustomElementController | HTMLElement;
@@ -38,7 +40,7 @@ export type RouteContextLike = IRouteContext | ICustomElementViewModel | ICustom
  * NOTE: differs from `Routeable` only in having `IViewportInstruction` instead of `IChildRouteConfig`
  * (which in turn are quite similar, but do have a few minor but important differences that make them non-interchangeable)
  */
-export type NavigationInstruction = string | IViewportInstruction | RouteableComponent;
+export type NavigationInstruction = string | IViewportInstruction | RouteableComponent | NavigationStrategy;
 
 /**
  * A component type, instance of definition that can be navigated to:
@@ -109,7 +111,7 @@ export class ViewportInstruction<TComponent extends ITypedNavigationInstruction_
         instruction.open ?? 0,
         instruction.close ?? 0,
         instruction.recognizedRoute ?? null,
-        component,
+        component as ITypedNavigationInstruction_Component,
         instruction.viewport ?? null,
         Object.freeze(instruction.params ?? null),
         children,
@@ -375,6 +377,13 @@ export class ViewportInstructionTree {
   }
 }
 
+type NavigationStrategyComponent = string | RouteType | Promise<IModule> | CustomElementDefinition;
+export class NavigationStrategy {
+  public constructor(
+    /** @internal */ public readonly getComponent: (viewportInstruction: IViewportInstruction, ctx: IRouteContext, node: RouteNode, route: RecognizedRoute<unknown>) => NavigationStrategyComponent,
+  ) { }
+}
+
 _START_CONST_ENUM();
 /**
  * @internal
@@ -385,6 +394,7 @@ export const enum NavigationInstructionType {
   CustomElementDefinition,
   Promise,
   IRouteViewModel,
+  NavigationStrategy,
 }
 _END_CONST_ENUM();
 
@@ -403,10 +413,12 @@ export interface ITypedNavigationInstruction_ViewportInstruction extends ITypedN
 export interface ITypedNavigationInstruction_CustomElementDefinition extends ITypedNavigationInstruction<CustomElementDefinition, NavigationInstructionType.CustomElementDefinition> { }
 export interface ITypedNavigationInstruction_Promise extends ITypedNavigationInstruction<Promise<IModule>, NavigationInstructionType.Promise> { }
 export interface ITypedNavigationInstruction_IRouteViewModel extends ITypedNavigationInstruction<IRouteViewModel, NavigationInstructionType.IRouteViewModel> { }
+export interface ITypedNavigationInstruction_NavigationStrategy extends ITypedNavigationInstruction<NavigationStrategy, NavigationInstructionType.NavigationStrategy> { }
 
 export type ITypedNavigationInstruction_T = (
   ITypedNavigationInstruction_Component |
-  ITypedNavigationInstruction_ViewportInstruction
+  ITypedNavigationInstruction_ViewportInstruction |
+  ITypedNavigationInstruction_NavigationStrategy
 );
 
 export type ITypedNavigationInstruction_Component = (
@@ -441,6 +453,7 @@ export class TypedNavigationInstruction<TInstruction extends NavigationInstructi
     if (typeof instruction === 'string') return new TypedNavigationInstruction(NavigationInstructionType.string, instruction);
     // Typings prevent this from happening, but guard it anyway due to `as any` and the sorts being a thing in userland code and tests.
     if (!isObjectOrFunction(instruction)) expectType('function/class or object', '', instruction);
+    if (instruction instanceof NavigationStrategy) return new TypedNavigationInstruction(NavigationInstructionType.NavigationStrategy, instruction);
     if (typeof instruction === 'function') {
       if (CustomElement.isType(instruction as Constructable)) {
         // This is the class itself
@@ -459,11 +472,30 @@ export class TypedNavigationInstruction<TInstruction extends NavigationInstructi
     if (isCustomElementViewModel(instruction)) return new TypedNavigationInstruction(NavigationInstructionType.IRouteViewModel, instruction);
     // We might have gotten a complete definition. In that case use it as-is.
     if (instruction instanceof CustomElementDefinition) return new TypedNavigationInstruction(NavigationInstructionType.CustomElementDefinition, instruction);
+
+    // If we have a partial definition, create a complete definition from it.
+    // Use-case:
+    // import * as component from './conventional-html-only-component.html';
+    // @route({
+    //   routes: [
+    //     {
+    //       path: 'path',
+    //       component,
+    //     },
+    //   ],
+    // })
+    if (isPartialCustomElementDefinition(instruction)) {
+      const definition = CustomElementDefinition.create(instruction);
+      CustomElement.define(definition);
+      return new TypedNavigationInstruction(NavigationInstructionType.CustomElementDefinition, definition);
+    }
+
     throw new Error(getMessage(Events.instrInvalid, tryStringify(instruction)));
   }
 
   public equals(this: ITypedNavigationInstruction_T, other: ITypedNavigationInstruction_T): boolean {
     switch (this.type) {
+      case NavigationInstructionType.NavigationStrategy:
       case NavigationInstructionType.CustomElementDefinition:
       case NavigationInstructionType.IRouteViewModel:
       case NavigationInstructionType.Promise:
@@ -488,6 +520,7 @@ export class TypedNavigationInstruction<TInstruction extends NavigationInstructi
         return this.value.name;
       case NavigationInstructionType.IRouteViewModel:
       case NavigationInstructionType.Promise:
+      case NavigationInstructionType.NavigationStrategy:
         throw new Error(getMessage(Events.instrInvalidUrlComponentOperation, this.type));
       case NavigationInstructionType.ViewportInstruction:
         return this.value.toUrlComponent();
@@ -501,6 +534,8 @@ export class TypedNavigationInstruction<TInstruction extends NavigationInstructi
     switch (this.type) {
       case NavigationInstructionType.CustomElementDefinition:
         return `CEDef(name:'${this.value.name}')`;
+      case NavigationInstructionType.NavigationStrategy:
+        return `NS`;
       case NavigationInstructionType.Promise:
         return `Promise`;
       case NavigationInstructionType.IRouteViewModel: