+++ packages/router-lite/src/route.ts
@@ -1,16 +1,16 @@
 import { Metadata } from '@aurelia/metadata';
 import { Constructable, emptyArray, onResolve, ResourceType, Writable, getResourceKeyFor } from '@aurelia/kernel';
 
-import { validateRouteConfig, expectType, shallowEquals, isPartialRedirectRouteConfig, isPartialChildRouteConfig } from './validation';
-import { defaultViewportName, ITypedNavigationInstruction_Component, NavigationInstructionType, TypedNavigationInstruction, ViewportInstruction } from './instructions';
+import { validateRouteConfig, expectType, shallowEquals, isPartialRedirectRouteConfig, isPartialChildRouteConfig, isPartialCustomElementDefinition } from './validation';
+import { defaultViewportName, ITypedNavigationInstruction_Component, ITypedNavigationInstruction_NavigationStrategy, IViewportInstruction, NavigationInstructionType, NavigationStrategy, TypedNavigationInstruction, ViewportInstruction } from './instructions';
 import type { RouteNode } from './route-tree';
 import type { IRouteContext } from './route-context';
-import { CustomElement, CustomElementDefinition } from '@aurelia/runtime-html';
+import { CustomElement, CustomElementDefinition, PartialCustomElementDefinition } from '@aurelia/runtime-html';
 import { IRouteViewModel } from './component-agent';
 import { ensureArrayOfStrings, ensureString } from './util';
 import type { FallbackFunction, IChildRouteConfig, IRedirectRouteConfig, IRouteConfig, Routeable, TransitionPlan, TransitionPlanOrFunc } from './options';
 import { Events, getMessage } from './events';
-import { RESIDUE } from '@aurelia/route-recognizer';
+import { RecognizedRoute, RESIDUE } from '@aurelia/route-recognizer';
 
 export const noRoutes = emptyArray as RouteConfig['routes'];
 
@@ -21,9 +21,15 @@ export class RouteConfig implements IRouteConfig, IChildRouteConfig {
   public get path(): string[] {
     const path = this._path;
     if (path.length > 0) return path;
-    const ceDfn = CustomElement.getDefinition(this.component as RouteType);
+    const ceDfn = CustomElement.getDefinition(this._component as RouteType);
     return this._path = [ceDfn.name, ...ceDfn.aliases];
   }
+  /** @internal */ private _currentComponent: Routeable | null = null;
+  public get component(): Routeable { return this._getComponent(); }
+
+  /** @internal */ public readonly _isNavigationStrategy: boolean;
+  /** @internal */ private readonly _component: Routeable | NavigationStrategy;
+
   private constructor(
     public readonly id: string,
     /** @internal */
@@ -36,9 +42,12 @@ export class RouteConfig implements IRouteConfig, IChildRouteConfig {
     public readonly data: Record<string, unknown>,
     public readonly routes: readonly Routeable[],
     public readonly fallback: Routeable | FallbackFunction | null,
-    public readonly component: Routeable,
+    component: Routeable | NavigationStrategy,
     public readonly nav: boolean,
-  ) { }
+  ) {
+    this._component = component;
+    this._isNavigationStrategy = component instanceof NavigationStrategy;
+  }
 
   /** @internal */
   public static _create(
@@ -130,7 +139,7 @@ export class RouteConfig implements IRouteConfig, IChildRouteConfig {
       config.data ?? this.data,
       config.routes ?? this.routes,
       config.fallback ?? this.fallback ?? parentConfig?.fallback ?? null,
-      this.component, // The RouteConfig is created using a definitive Type as component; do not overwrite it.
+      this._component, // The RouteConfig is created using a definitive Type as component; do not overwrite it.
       config.nav ?? this.nav,
     );
   }
@@ -198,7 +207,7 @@ export class RouteConfig implements IRouteConfig, IChildRouteConfig {
       this.data,
       this.routes,
       this.fallback,
-      this.component,
+      this._component,
       this.nav,
     );
   }
@@ -210,6 +219,47 @@ export class RouteConfig implements IRouteConfig, IChildRouteConfig {
       && !CustomElement.isType(fallback as Constructable)
       ? (fallback as FallbackFunction)(viewportInstruction, routeNode, context)
       : fallback;
+  }
+
+  /** @internal */
+  public _getComponentName(): string {
+    try {
+      return (this._getComponent() as RouteType).name;
+    } catch {
+      // TODO(Sayan): Convert all Errors in router lite to instances of RouterError
+      return 'UNRESOLVED-NAVIGATION-STRATEGY';
+    }
+  }
+
+  /** @internal */
+  public _getComponent(): Routeable;
+  public _getComponent(vi: IViewportInstruction, ctx: IRouteContext, node: RouteNode, route: RecognizedRoute<unknown>): Routeable;
+  public _getComponent(vi?: IViewportInstruction, ctx?: IRouteContext, node?: RouteNode, route?: RecognizedRoute<unknown>): Routeable {
+    if (vi == null) {
+      if (this._currentComponent != null) return this._currentComponent;
+      if (this._isNavigationStrategy) throw new Error(getMessage(Events.rtInvalidOperationNavigationStrategyComponent, this.id));
+      return this._currentComponent = this._component;
+    }
+    return this._currentComponent ??= this._isNavigationStrategy ? (this._component as NavigationStrategy).getComponent(vi, ctx!, node!, route!) : this._component;
+  }
+
+  /** @internal */
+  public _handleNavigationStart(): void {
+    if (!this._isNavigationStrategy) return;
+    this._currentComponent = null;
+  }
+
+  public toString(): string {
+    let value = `RConf(id: ${this.id}, isNavigationStrategy: ${this._isNavigationStrategy}`;
+    if (!__DEV__) return `{${value}})`;
+
+    value += `, path: [${this.path.join(',')}]`;
+    if (this.redirectTo) value += `, redirectTo: ${this.redirectTo}`;
+    if (this.caseSensitive) value += `, caseSensitive: ${this.caseSensitive}`;
+    if (this.transitionPlan != null) value += `, transitionPlan: ${this.transitionPlan}`;
+    value += `, viewport: ${this.viewport}`;
+    if (this._currentComponent != null) value += `, component: ${(this._currentComponent as RouteType).name}`;
+    return `${value})`;
   }
 }
 
@@ -286,12 +336,14 @@ export function resolveRouteConfiguration(routeable: Routeable, isChild: boolean
   if (isPartialRedirectRouteConfig(routeable)) return RouteConfig._create(routeable, null);
 
   const [instruction, ceDef] = resolveCustomElementDefinition(routeable, context);
+  if (instruction.type === NavigationInstructionType.NavigationStrategy)
+    return RouteConfig._create({ ...routeable as IChildRouteConfig, nav: false }, null);
 
-  return onResolve(ceDef, $ceDef => {
+  return onResolve(ceDef!, $ceDef => {
     const type = $ceDef.Type;
     const routeConfig = Route.getConfig(type);
 
-    // If the component is used as a child, then apply the child configuration (comping from parent) and return a new RouteConfig with the configuration applied.
+    // If the component is used as a child, then apply the child configuration (coming from parent) and return a new RouteConfig with the configuration applied.
     if (isPartialChildRouteConfig(routeable)) return routeConfig._applyChildRouteConfig(routeable, parent);
 
     // If the component is used as a child, then return a clone.
@@ -310,17 +362,26 @@ export function resolveRouteConfiguration(routeable: Routeable, isChild: boolean
 }
 
 /** @internal */
-export function resolveCustomElementDefinition(routeable: Routeable, context: IRouteContext | null | undefined): [instruction: ITypedNavigationInstruction_Component, ceDef: CustomElementDefinition | Promise<CustomElementDefinition>] {
+export function resolveCustomElementDefinition(routeable: Routeable, context: IRouteContext | null | undefined):
+  | [instruction: ITypedNavigationInstruction_Component, ceDef: CustomElementDefinition | Promise<CustomElementDefinition>]
+  | [instruction: ITypedNavigationInstruction_NavigationStrategy, ceDef: null] {
   const instruction = createNavigationInstruction(routeable);
   let ceDef: CustomElementDefinition | Promise<CustomElementDefinition>;
   switch (instruction.type) {
+    case NavigationInstructionType.NavigationStrategy: return [instruction, null];
     case NavigationInstructionType.string: {
       if (context == null) throw new Error(getMessage(Events.rtNoCtxStrComponent));
 
-      const component = CustomElement.find(context.container, instruction.value);
+      const dependencies = context.component.dependencies;
+      let component = dependencies.find(d => isPartialCustomElementDefinition(d) && d.name === instruction.value) as CustomElementDefinition | PartialCustomElementDefinition
+        ?? CustomElement.find(context.container, instruction.value);
       if (component === null) throw new Error(getMessage(Events.rtNoComponent, instruction.value, context));
 
-      ceDef = component;
+      if (!(component instanceof CustomElementDefinition)) {
+        component = CustomElementDefinition.create(component);
+        CustomElement.define(component);
+      }
+      ceDef = component as CustomElementDefinition;
       break;
     }
     case NavigationInstructionType.CustomElementDefinition:
@@ -335,10 +396,10 @@ export function resolveCustomElementDefinition(routeable: Routeable, context: IR
       ceDef = context._resolveLazy(instruction.value);
       break;
   }
-  return [instruction, ceDef];
+  return [instruction, ceDef!];
 }
 
-function createNavigationInstruction(routeable: Exclude<Routeable, IRedirectRouteConfig>): ITypedNavigationInstruction_Component {
+function createNavigationInstruction(routeable: Exclude<Routeable, IRedirectRouteConfig>): ITypedNavigationInstruction_Component | ITypedNavigationInstruction_NavigationStrategy {
   return isPartialChildRouteConfig(routeable)
     ? createNavigationInstruction(routeable.component)
     : TypedNavigationInstruction.create(routeable);