+++ packages/runtime-html/src/observation/value-attribute-observer.ts
@@ -4,7 +4,7 @@ import { mixinNodeObserverUseConfig } from './observation-utils';
 import { atLayout, atNode, atObserver } from '../utilities';
 
 import type { AccessorType, ISubscriberCollection } from '@aurelia/runtime';
-import type { INode } from '../dom';
+import type { INode } from '../dom.node';
 import type { INodeObserver, INodeObserverConfigBase } from './observer-locator';
 
 export interface ValueAttributeObserver extends ISubscriberCollection {}