+++ packages/validation-html/src/validation-controller.ts
@@ -14,6 +14,7 @@ import {
 } from '@aurelia/expression-parser';
 import {
   astEvaluate,
+  IConnectable,
   type Scope,
 } from '@aurelia/runtime';
 import {
@@ -97,6 +98,7 @@ export interface ValidationResultsSubscriber {
  */
 export class BindingInfo {
   /**
+   * @param {IConnectable} sourceObserver - An observer for the binding source.
    * @param {Element} target - The HTMLElement associated with the binding.
    * @param {Scope} scope - The binding scope.
    * @param {PropertyRule[]} [rules] - Rules bound to the binding behavior.
@@ -104,6 +106,7 @@ export class BindingInfo {
    * @memberof BindingInfo
    */
   public constructor(
+    public sourceObserver: IConnectable,
     public target: Element,
     public scope: Scope,
     public rules?: PropertyRule[],
@@ -145,7 +148,7 @@ export function getPropertyInfo(binding: BindingWithBehavior, info: BindingInfo)
           toCachePropertyName = keyExpr.$kind === 'PrimitiveLiteral';
         }
         // eslint-disable-next-line
-        memberName = `[${(astEvaluate(keyExpr, scope, binding, null) as any).toString()}]`;
+        memberName = `[${(astEvaluate(keyExpr, scope, binding, info.sourceObserver) as any).toString()}]`;
         break;
       }
       default:
@@ -164,7 +167,7 @@ export function getPropertyInfo(binding: BindingWithBehavior, info: BindingInfo)
     propertyName = expression.name;
     object = scope.bindingContext;
   } else {
-    object = astEvaluate(expression, scope, binding, null);
+    object = astEvaluate(expression, scope, binding, info.sourceObserver);
   }
   if (object === null || object === void 0) {
     return (void 0);